Additional work for creating and searching for users
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@5272 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/ui/ListSlider.java
@@ -480,7 +480,7 @@ public class ListSlider extends JPanel
         int [] indices = new int [indicesToMove.size()];
         for(int i = 0; i< indicesToMove.size(); i ++)
         {
-            indices[i] = (Integer)indicesToMove.get(i).intValue();    
+            indices[i] = indicesToMove.get(i).intValue();    
             log.debug("selecting: " +  (Integer)indicesToMove.get(i).intValue() );
         }
         sourceList.setSelectedIndices(indices);
